product test refactor packages refactor
