+++ src/main/java/com/ecommerce/eshop/product/controllers/IndexController.java
@@ -1,6 +1,5 @@
-package com.ecommerce.eshop.controllers;
+package com.ecommerce.eshop.product.controllers;
 
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
