+++ src/main/java/com/ecommerce/eshop/product/exceptions/CategoryNotFoundException.java
@@ -1,4 +1,4 @@
-package com.ecommerce.eshop.utils.excepctions;
+package com.ecommerce.eshop.product.exceptions;
 
 public class CategoryNotFoundException extends RuntimeException {
     public CategoryNotFoundException(String message) {
