+++ src/main/java/com/ecommerce/eshop/product/exceptions/ProductCreationException.java
@@ -1,4 +1,4 @@
-package com.ecommerce.eshop.utils.excepctions;
+package com.ecommerce.eshop.product.exceptions;
 
 public class ProductCreationException extends RuntimeException{
     public ProductCreationException(String message) {
