+++ src/main/java/com/ecommerce/eshop/product/exceptions/ProductNotFoundException.java
@@ -1,4 +1,4 @@
-package com.ecommerce.eshop.utils.excepctions;
+package com.ecommerce.eshop.product.exceptions;
 
 public class ProductNotFoundException extends RuntimeException{
     public ProductNotFoundException(String message) {
