+++ src/main/java/com/ecommerce/eshop/product/models/Product.java
@@ -1,4 +1,4 @@
-package com.ecommerce.eshop.models.product;
+package com.ecommerce.eshop.product.models;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
