+++ src/main/java/com/ecommerce/eshop/product/repositories/ImageRepository.java
@@ -1,6 +1,6 @@
-package com.ecommerce.eshop.repositories;
+package com.ecommerce.eshop.product.repositories;
 
-import com.ecommerce.eshop.models.product.ProductImage;
+import com.ecommerce.eshop.product.models.ProductImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
