+++ src/main/java/com/ecommerce/eshop/product/repositories/ProductRepository.java
@@ -1,6 +1,6 @@
-package com.ecommerce.eshop.repositories;
+package com.ecommerce.eshop.product.repositories;
 
-import com.ecommerce.eshop.models.product.Product;
+import com.ecommerce.eshop.product.models.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
